+++ src/lib/Data/TokenStream.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Archive/Token.hpp>
+#include "Token.hpp"
 
 namespace sedmgr {
