+++ src/lib/MockDevice/MockSession.hpp
@@ -20,18 +20,18 @@ public:
     void Output(std::span<std::byte> data);
 
 private:
-    void SessionManagerInput(const Method& method);
-    void SessionInput(Uid invokingId, const Method& method);
+    void SessionManagerInput(const MethodCall& method);
+    void SessionInput(Uid invokingId, const MethodCall& method);
 
-    void StartSession(const Method& method);
-    void Properties(const Method& method);
+    void StartSession(const MethodCall& method);
+    void Properties(const MethodCall& method);
     void EndSession();
 
-    void Next(MockSecurityProvider& sp, Uid invokingId, const Method& method);
-    void Get(MockSecurityProvider& sp, Uid invokingId, const Method& method);
-    void Set(MockSecurityProvider& sp, Uid invokingId, const Method& method);
+    void Next(MockSecurityProvider& sp, Uid invokingId, const MethodCall& method);
+    void Get(MockSecurityProvider& sp, Uid invokingId, const MethodCall& method);
+    void Set(MockSecurityProvider& sp, Uid invokingId, const MethodCall& method);
 
-    void EnqueueMethod(Uid invokingId, const Method& method);
+    void EnqueueMethod(Uid invokingId, const MethodCall& method);
     void EnqueueMethodResult(const MethodResult& result);
     void EnqueueResponse(const Value& value);
