+++ src/lib/TrustedPeripheral/Method.hpp
@@ -2,53 +2,12 @@
 
 #include <Archive/Types/ValueToNative.hpp>
 #include <Data/NativeTypes.hpp>
-#include <Data/Value.hpp>
+#include <Data/Method.hpp>
 
 #include <algorithm>
 
 
 namespace sedmgr {
-
-enum class eMethodStatus : uint8_t {
-    SUCCESS = 0x00,
-    NOT_AUTHORIZED = 0x01,
-    OBSOLETE = 0x02,
-    SP_BUSY = 0x03,
-    SP_FAILED = 0x04,
-    SP_DISABLED = 0x05,
-    SP_FROZEN = 0x06,
-    NO_SESSIONS_AVAILABLE = 0x07,
-    UNIQUENESS_CONFLICT = 0x08,
-    INSUFFICIENT_SPACE = 0x09,
-    INSUFFICIENT_ROWS = 0x0A,
-    INVALID_PARAMETER = 0x0C,
-    OBSOLETE_1 = 0x0D,
-    OBSOLETE_2 = 0x0E,
-    TPER_MALFUNCTION = 0x0F,
-    TRANSACTION_FAILURE = 0x10,
-    RESPONSE_OVERFLOW = 0x11,
-    AUTHORITY_LOCKED_OUT = 0x12,
-    FAIL = 0x3F,
-};
-
-
-struct Method {
-    Uid methodId;
-    std::vector<Value> args;
-    eMethodStatus status = eMethodStatus::SUCCESS;
-};
-
-
-struct MethodResult {
-    std::vector<Value> values;
-    eMethodStatus status = eMethodStatus::SUCCESS;
-};
-
-Value MethodToValue(Uid invokingId, const Method& method);
-Method MethodFromValue(const Value& stream);
-MethodResult MethodResultFromValue(const Value& stream);
-void MethodStatusToException(std::string_view methodName, eMethodStatus status);
-
 
 namespace impl {
